third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-type-to-search.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Type to search should focus but not select an option until selection is confirmed.
  assert_equals: The `bravo` option should be focused after pressing `b`. expected Element node <option id="bravo">bravo</option> but got Element node <select id="target">\n  <option></option>\n  <option>alpha<...
Harness: the test ran to completion.
